% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph_params.R
\name{build_graph_params}
\alias{build_graph_params}
\title{plot_ggraph}
\usage{
build_graph_params(input_igraph, color, width, scale = "diverge",
  pal = NULL)
}
\arguments{
\item{input_igraph}{the igraph object}

\item{color}{color of edges}

\item{width}{width of edges}

\item{scale}{"continuous" or "diverge"}

\item{pal}{Rcolorbrewer pal name. Pick appropriately for diverging or continuous data}
}
\description{
This function will assign color and width values to your igraph network
}
\details{
#
}
\examples{

}
