% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network}
\alias{build_network}
\title{build_network}
\usage{
build_network(input, from, to, color, weight)
}
\arguments{
\item{input}{the input dataframe}
}
\description{
This function will build a igraph network from a dataframe and calculate the layout.
}
\details{
#
}
\examples{

}
