% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vis_subnetwork.R
\name{build_vis_subnetwork}
\alias{build_vis_subnetwork}
\title{build_visnetwork}
\usage{
build_vis_subnetwork(input_igraph)
}
\arguments{
\item{input_igraph}{the igraph object}
}
\description{
This function will build a visnetwork object from and igraph object.
}
\details{
#
}
\examples{

}
