% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_visnet.R
\name{build_visnetwork}
\alias{build_visnetwork}
\title{build_visnetwork}
\usage{
build_visnetwork(input_igraph, graph_params = NULL, edge_scale = 7)
}
\arguments{
\item{input_igraph}{the ggraph plot}
}
\description{
This function will build a visnetwork object
}
\details{
#
}
\examples{

}
