% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggraph.R
\name{plot_ggraph}
\alias{plot_ggraph}
\title{plot_ggraph}
\usage{
plot_ggraph(
  input_igraph,
  color,
  width,
  col_lab = NULL,
  width_lab = NULL,
  scale = "diverge",
  render = T,
  colpal = c("blue", "red"),
  label = F,
  label_nodes = NULL
)
}
\arguments{
\item{input_igraph}{the igraph object}

\item{color}{color of edges}

\item{width}{width of edges}

\item{col_lab}{label of color legend}

\item{width_lab}{label of width legend}

\item{scale}{"continuous" or "diverge"}

\item{render}{Whether or not a plot will be rendered}

\item{colpal}{a vector of length 2 for colors}

\item{label}{whether or not to label nodes}

\item{label_nodes}{subset of nodes to label}
}
\description{
This function will plot a ggraph object from an igraph object
}
\details{

}
\examples{
test_net <- plot_ggraph(input_igraph = test_net, color = "rho", width = "log_p", scale = "diverge")
}
