% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_net.R
\name{shiny_net}
\alias{shiny_net}
\title{Shiny_net}
\usage{
shiny_net(input)
}
\arguments{
\item{input}{the ggraph plot}
}
\description{
This function will take the visnetwork and start Shiny
}
\details{
#
}
\examples{

}
