% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_net}
\alias{test_net}
\title{Gene correlation network}
\format{
An object of class \code{data.frame} with 1118 rows and 5 columns.
}
\usage{
test_net
}
\description{
Gene correlation network
}
\keyword{datasets}
