% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_visnetwork.R
\name{build_visnetwork}
\alias{build_visnetwork}
\title{build_visnetwork}
\usage{
build_visnetwork(input_igraph)
}
\arguments{
\item{input_igraph}{the igraph object}
}
\description{
This function will build a visnetwork object from and igraph object.
}
\details{
#
}
\examples{
build_visnetwork(input_igraph = test_net)
}
