% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ggraph.R
\name{extract_ggraph}
\alias{extract_ggraph}
\title{extract_ggraph}
\usage{
extract_ggraph(input_igraph, input_ggraph)
}
\arguments{
\item{input_igraph}{the igraph object. ggraph parameters will be written in}

\item{input_ggraph}{the ggraph object.}
}
\description{
This function will extract plotting parameters from the ggraph object
}
\details{
#
}
\examples{

}
