% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ggraph.R
\name{extract_ggraph}
\alias{extract_ggraph}
\title{extract_ggraph}
\usage{
extract_ggraph(input)
}
\arguments{
\item{input}{the ggraph plot}
}
\description{
This function will extract plotting parameters from the ggraph object
}
\details{
#
}
\examples{

}
