% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_visnet.R
\name{make_visnet}
\alias{make_visnet}
\title{make_visnet}
\usage{
make_visnet(input_igraph, edge_scale = 7)
}
\arguments{
\item{input_igraph}{the igraph object}

\item{edge_scale}{Default = 7. Scale factor for difference between igraph and visnet edge sizes.}
}
\description{
This function will build a visnetwork object from and igraph object
}
\details{
#
}
\examples{

}
