% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggraph.R
\name{plot_ggraph}
\alias{plot_ggraph}
\title{build_network}
\usage{
plot_ggraph(input, color, width, col_lab = NULL, width_lab = NULL,
  scale = "diverge", extract = T)
}
\arguments{
\item{input}{the input dataframe}
}
\description{
This function will build a igraph network from a dataframe and calculate the layout.
}
\details{
#
}
\examples{

}
